requirements.txt
psycopg2-binary
fastapi
uvicorn
aiogram
sqlalchemy
requests
alembic
asyncpg
greenlet
httpx
sqlalchemy
sqlalchemy[asyncio]
asyncpg
pyjwt
pydantic
python-jose
python-multipart
passlib[bcrypt]
pydantic-settings
pydantic[email]
fastapi_jwt
starlette
python-dotenv
